Rename combobox components to kebab-case
