--- src/components/Combobox/index.ts
@@ -1,7 +1,0 @@
-export * from './Combobox';
-export * from './ComboboxContent';
-export * from './ComboboxEmpty';
-export * from './ComboboxInput';
-export * from './ComboboxItem';
-export * from './context';
-export * from './types';
